refactor(PN-11896): refactor PG for digital courtesy contacts
