+++ packages/pn-personagiuridica-webapp/src/components/Contacts/CourtesyContacts.tsx
@@ -1,10 +1,11 @@
 import { useTranslation } from 'react-i18next';
 
-import { Alert, Box, Typography } from '@mui/material';
+import { Alert, Box, Stack, Typography } from '@mui/material';
 import { IllusEmail } from '@pagopa/mui-italia';
 
-import { DigitalAddress } from '../../models/contacts';
-import CourtesyContactsList from './CourtesyContactsList';
+import { CourtesyChannelType, DigitalAddress } from '../../models/contacts';
+import { countContactsByType } from '../../utility/contacts.utility';
+import CourtesyContactItem, { CourtesyFieldType } from './CourtesyContactItem';
 import DigitalContactsCard from './DigitalContactsCard';
 
 interface Props {
@@ -13,6 +14,19 @@ interface Props {
 
 const CourtesyContacts: React.FC<Props> = ({ contacts }) => {
   const { t } = useTranslation(['common', 'recapiti']);
+
+  const phoneContact = contacts.find(
+    (contact) => contact.channelType === CourtesyChannelType.SMS && contact.senderId === 'default'
+  );
+  const emailContact = contacts.find(
+    (contact) => contact.channelType === CourtesyChannelType.EMAIL && contact.senderId === 'default'
+  );
+
+  const phoneContactsQuantity = (): number =>
+    countContactsByType(contacts, CourtesyChannelType.SMS);
+
+  const emailContactsQuantity = (): number =>
+    countContactsByType(contacts, CourtesyChannelType.EMAIL);
 
   return (
     <DigitalContactsCard
@@ -25,7 +39,18 @@ const CourtesyContacts: React.FC<Props> = ({ contacts }) => {
       avatar={<IllusEmail size={60} />}
     >
       <Box sx={{ width: { xs: '100%', lg: '50%' } }} data-testid="courtesyContacts">
-        <CourtesyContactsList contacts={contacts} />
+        <Stack spacing={3} mt={3}>
+          <CourtesyContactItem
+            type={CourtesyFieldType.EMAIL}
+            value={emailContact?.value ? emailContact.value : ''}
+            blockDelete={emailContactsQuantity() > 1}
+          />
+          <CourtesyContactItem
+            type={CourtesyFieldType.PHONE}
+            value={phoneContact?.value ? phoneContact.value : ''}
+            blockDelete={phoneContactsQuantity() > 1}
+          />
+        </Stack>
       </Box>
       <Alert
         role="banner"
