+++ packages/pn-personagiuridica-webapp/src/components/Contacts/__test__/DigitalContactsCard.test.tsx
@@ -5,7 +5,6 @@ import DigitalContactsCard from '../DigitalContactsCard';
 
 const title = 'Mocked title';
 const subTitle = 'Mocked subtitle';
-const actions = <button>Click me</button>;
 const body = <div data-testid="body">Body</div>;
 
 vi.mock('react-i18next', () => ({
@@ -16,32 +15,7 @@ vi.mock('react-i18next', () => ({
 }));
 
 describe('DigitalContactsCard Component', () => {
-  it('renders component - with actions', () => {
-    // render component
-    const result = render(
-      <DigitalContactsCard
-        sectionTitle={'mocked-sectionTitle'}
-        title={title}
-        subtitle={subTitle}
-        actions={actions}
-        avatar="avatar"
-      >
-        {body}
-      </DigitalContactsCard>
-    );
-    const titleEl = result.container.querySelector('h6');
-    expect(titleEl).toBeInTheDocument();
-    expect(titleEl).toHaveTextContent(title);
-    expect(result.container).toHaveTextContent(subTitle);
-    const bodyEl = result.getByTestId('body');
-    expect(bodyEl).toBeInTheDocument();
-    expect(bodyEl).toHaveTextContent(/Body/i);
-    const buttonEl = result.container.querySelector('button');
-    expect(buttonEl).toBeInTheDocument();
-    expect(buttonEl).toHaveTextContent(/Click me/i);
-  });
-
-  it('renders component - without actions', () => {
+  it('renders component', () => {
     // render component
     const result = render(
       <DigitalContactsCard
